Avoid test-failure without jar
When we build the package in OBS or osc build, we run make check
and there is no jar available. This should not fail.

+++ tests/jar-test.sh
@@ -3,6 +3,7 @@
 describe "roundup(1) testing of jar"
 
 it_normalizes_mtime_on_create() {
+   test -e /usr/bin/jar || return 0
    mkdir -p tmp
    echo foo > tmp/foo
    SOURCE_DATE_EPOCH=1 ../bin/jar --create --file tmp/test.jar tmp/foo
